dollar rate

+++ app/src/main/java/uz/texnopos/elektrolife/data/model/newproduct/Amount.kt
@@ -1,7 +1,7 @@
 package uz.texnopos.elektrolife.data.model.newproduct
 
 data class Amount(
-    var wholesalePrice: Long = 0L,
+    var wholesalePrice: Double = 0.0,
     var minPrice: Long = 0L,
     var maxPrice: Long = 0L
 )
